+++ classes/mage.lua
@@ -246,6 +246,7 @@ local FrozenHandler = {
     deep_freeze = { 44572 }, -- Deep Freeze is both a debuff for 'Frozen' Spell Alert and its own Glowing Button
     deep_freeze_sod = { 428739 }, -- Season of Discovery
     hungering_cold = { 49203 }, -- from Death Knights
+    improved_cone_of_cold = { 83301, 83302 },
     ring_of_frost = { 82691 },
 
     freezeID = 5276, -- Not really a 'Frozen' spell ID, but the name should help players identify the intent
@@ -279,6 +280,7 @@ local FrozenHandler = {
         self:addSpellIDCandidates(self.deep_freeze);
         self:addSpellIDCandidates(self.deep_freeze_sod);
         self:addSpellIDCandidates(self.hungering_cold);
+        self:addSpellIDCandidates(self.improved_cone_of_cold);
         self:addSpellIDCandidates(self.ring_of_frost);
 
         self.freezable = self:isTargetFreezable();